ublas::gemm support for parallelised matrix multiplications
This patch adds parallel version of gemm for openmp version 4.0 and above.
Two versions are provided.  One that is supposed to scale to the cores of
a single socket, and another that is supposed to scale to multiple sockets
(but fails to do so, when compiled by icc)

The version that scales through sockets needs nested parallelism enabled.

Signed-off-by: Imre Palik <imre_palik@yahoo.co.uk>